feat: add team support
