+++ apps/dashboard/src/components/buttons/team/change-team-button.tsx
@@ -1,11 +1,9 @@
 "use client";
 
-import { ChangeTeamModal } from "@/components/modals/team/change-team-modal";
-import { TeamSchema } from "@midday/supabase/types";
+import { ChangeTeamSheet } from "@/components/modals/team/change-team-sheet";
 import { Button } from "@midday/ui/button";
-import { SwitchIcon } from "@radix-ui/react-icons";
 import { LoaderPinwheel } from "lucide-react";
-import React, { useState } from "react";
+import { useState } from "react";
 
 interface ChangeTeamButtonProps {
   currentTeamId: any;
@@ -26,9 +24,9 @@ export function ChangeTeamButton({
     <>
       <Button onClick={handleOpenModal}>
         <LoaderPinwheel className="w-6 h-6 mr-2" />
-        Change Team
+        Team
       </Button>
-      <ChangeTeamModal
+      <ChangeTeamSheet
         isOpen={isOpen}
         onOpenChange={setIsOpen}
         teams={teams}
