+++ apps/dashboard/src/components/team-menu.tsx
@@ -1,5 +1,8 @@
 import { TeamDropdown } from "@/components/team-dropdown";
 import { getTeams, getUser } from "@midday/supabase/cached-queries";
+import { DropdownMenu, DropdownMenuContent, DropdownMenuGroup, DropdownMenuItem, DropdownMenuShortcut, DropdownMenuTrigger } from "@midday/ui/dropdown-menu";
+import { LoaderPinwheel } from "lucide-react";
+import Link from "next/link";
 import { ChangeTeamButton } from "./buttons/team/change-team-button";
 
 interface TeamMenuProps {
